+++ scripts/plot_dat.py
@@ -1,11 +1,11 @@
-from dat import *
-
+import argparse
 import os
 import subprocess
+from typing import List
 import numpy as np
 import plot
 
-ESMINI_PATH = os.path.realpath(os.pardir)
+ESMINI_PATH = os.path.realpath(os.getcwd())
 
 def get_labels_line_extended() -> List[str]:
     """ Get the extended labels line """
@@ -28,8 +28,8 @@ if __name__ == "__main__":
     # Execute the parse_args() method
     args = parser.parse_args()
 
-    esmini_args = [os.path.join(ESMINI_PATH,'bin','dat2csv'), args.filename, '--extended', '--quiet']
-    process = subprocess.Popen(esmini_args, cwd=os.path.dirname(os.path.realpath(__file__)),stdout=subprocess.PIPE, text=True)
+    esmini_args = [os.path.join(ESMINI_PATH,'bin','dat2csv'), "--file", args.filename, '--extended', '--print_csv']
+    process = subprocess.Popen(esmini_args, cwd=ESMINI_PATH, stdout=subprocess.PIPE, text=True)
     csv_data = process.stdout.read().strip().splitlines()
 
     data = np.genfromtxt(csv_data[1:], delimiter=',', names=True, dtype=None, encoding=None, autostrip=True)